find and run doctests.
had to go through quite a bit to get only doctests in wardenclyffe discovered and run.
by default django-nose wants to run tests that it finds in virtualenv.py and pip.py
(which fail *sigh*), plus everything it finds in the virtualenv (we do *not* want
to run celery's test suite each time) and migrations directories. so i installed
nose-exclude and configured it to ignore those directories. plus with the --noexe
option, and setting virtualenv.py and pip.py executable, it can avoid those