audio mp3 -> mp4 conversion. PMT #97710
This changes the workflow for mediathread uploaded audio files to have them encoded locally,
then sent up to S3 and Elastic Transcoder as video files. This completely eliminates
Podcast Producer from Wardenclyffe.

This feature is behind a feature flag called 'encode_audio', which is currently turned off
in production. Once this code is deployed, we can turn that flag on for testing, verify that
the pipeline all works, then turn the flag on for the general public and start working on
removing the PCP specific code from Wardenclyffe.