:bug: if the video is uploaded to the public collection, put it in the public directory. PMT #97823
Specifically, when it comes back from ET, we need to check if it's in the
public collection, and if so, upload it to the public directory
on the CUIT media server.

For now, it's basically hard-coded as to which collection is public
(defaulting to secure for anything else). In lieu of a more comprehensive
approach to configuring where things end up, this will have to do.