remove PCP/Kino smoketest
In theory, WC still uses PCP, but only on 'internal' functionality, ie, operations
that must be explicitly initiated by video team staff. It's no longer in use
for public facing functionality (mediathread uploads) and is on the slate
for complete removal in the coming months.

The PCP server has been so unreliable lately that I'm now ready to just pull the smoketest out
of WC. Ie, if PCP's down, I just don't care that much anymore.