automatically retry operations
Especially with the AWS stuff, sometimes operations just timeout
for reasons that are out of our control. We have manual 'rerun' buttons
and the video team knows to use them. But Celery supports automatically
retrying tasks, so let's see if Wardenclyffe can just do that for them.

I implemented it with an exponential backoff and it maxes out at 10
retries (2^10 seconds is about 17 minutes).