graphite metrics for retries
It would be nice to be able to keep track of how often tasks get
retried, now many retries happen, and whether they are maxing out.

This adds a couple metrics to graphite/statsd for that. There's one
counter that for every retry that happens, one for when max_retries is
reached, and then individual 'retry_01', 'retry_02', etc counters so we
can get a sense of how many retries it takes to get an operation to
complete. I think these will be handy to graph and alert off of.