handle simultaneous SNS responses a little better
This sentry error:
https://sentry.ccnmtl.columbia.edu/sentry-internal/wardenclyffe/group/911/

seems to come from amazon occasionally sending duplicate SNS events and
triggering a race condition. (this can happen on this view since it's in
non-atomic request mode).

It's not frequent and doesn't break anything, and this doesn't totally
fix the issue, but it reduces the span between checking if the specified
`File` exists and working with it, so the race condition should at least
be less common.