direct to s3 upload form
This is in parallel to the regular upload form so we can start testing
it in production against the real AWS/EC2 setup.

There are still some big missing pieces like audio workflow and upload
to youtube, but this should be enough that we can sanity check the
general approach.