make a source file object on s3 upload
I'd taken that out of the straight to s3 workflow reasoning that we no
longer had a source file on the filesystem.

It looks like the extract metadata task still expects it to exist though
so it can stash metadata on it.

Easy enough to put it back in.

Once we've moved everything to S3 upload only, it will probably be much
simpler to just have the extract metadata job create it itself instead
of relying on it to already have been created.