fix audio workflow for mediathread s3 upload
OK, it turns out the audio workflow on a mediathread
upload is slightly different than a regular audio upload.

Instead of directly queueing up the tasks that would normally happen
after the S3 upload completes (extract metadata, and create ET job), we
instead trigger the audio encode (the "local audio encode" task is
poorly named, it does actually pull from S3 first). When that completes
it will trigger an S3 upload (of the generated mp4), which in turn will
trigger the normal tasks.

So when we got an audio upload for mediathread, we were queueing up
those other tasks immediately, with the .mp3 file as the
source. Instead, they need to wait until there's an .mp4 ready.