limit file uploads to video/audio
on the mediathread upload form, we actually know if they're supposed to
be uploading a video or an audio file and can limit accordingly. on the
regular upload, we don't necessarily know until we know which collection
they are uploading to, which we have no way of knowing yet, so we have
to allow both video and audio types.