add model for tracking drop buckets
This is laying the foundations for having S3 "dropbox" buckets.

The eventual goal is that the video team and/or faculty/clients can
upload files directly to an S3 bucket, using a nice desktop tool and
have that automatically kick off the full pipeline without any further
manual steps.

S3 lets you register SNS and SQS notifications for various events,
including 's3:ObjectCreated`, which would be triggered when a file
upload completes. So that should be doable. We can create buckets,
configure IAM roles and access policies however we like (public write
only or handing out credentials to whomever would be uploading,
etc). Then we can configure the buckets to send these notifications to
an SNS topic which we then point at an endpoint in WC, which moves the
uploaded video into the "main" input bucket and kicks off the pipeline
from there like normal.

The regular upload process collects a few other bits of metadata though
that are required for everything else to work. Eg, when a video is
uploaded, we need to associate it with a Collection, which in turn
dictates whether audio conversion happens or whether the video will
eventually be sent to the public or secure servers. We also always
associate a user (by UNI) with the video so we know who uploaded it and
so we can email them once the transcoding and uploading has completed or
in the case of a failure. With an S3 dropbox, we don't have any of that
metadata available to us.

This commit adds a 'DropBucket' model which we can use to set up
mappings from S3 buckets to a Collection and a User. This will let the
SNS endpoint send the video into the correct collection
automatically (the 's3:ObjectCreated' notification should contain the
bucket id, so WC can use that as a key to look up the mappings).

For now, I imagine we would just set these up manually through Django
Admin. That's not much extra admin work compared to setting up the S3
bucket, access policies, etc on the AWS side. Eventually, when we have a
better handle on how those things typically get configured, we might
build a self-serve interface to let the video team set up a new drop
bucket and configure it themselves. But that's probably a ways off.