dropbox SNS debug view
This is purely for information gathering. When SNS sends a message to
this endpoint, it just reads in the request and emails it to me so I can
see what they are sending.

The eventual goal is to create an SNS topic that passes along events
when files are uploaded to the S3 drop bucket, then to subscribe this
URL to that SNS topic.

What I discovered previously, doing SNS subscriptions on the elastic
transcoder workflow was that the AWS documentation is thin and confusing
when it comes to the exact format of the message payload that they send
you. The simplest approach is just to put something like this up and
trigger a few events and just see what you get back and figure out from
there what headers and message fields are actually useful for
identifying and routing.

So this whole view as it exists is intended to be disposable. Once it's
deployed and I hit it with some test data, it will be replaced with the
beginnings of the actual endpoint that kicks off a pipeline, etc.