declare bankruptcy on petiole
petiole was my old design for "next-gen" wardenclyffe workflow
management. I started implementing the model long ago, but never got it
to the point where it was ready to start actually integrating.

In the meantime, the workflow system in WC has evolved on its own, the
problem that it addresses has changed significantly (no more PCP; AWS
integration instead), and our understanding of what the system needs to
handle has changed.

I've been working on a new design that has some similar goals to
petiole (letting users/admins define new workflows in a flexible
manner), but is far enough from petiole that it's easier to just wipe
this and start clean.