add s3serve view
Beginning to play around with pointing a player straight at the encoded
video on S3.

S3 does "pseudo-streaming" with range headers basically the same as our
FLV solution does, so this could potentially be "good enough" from a
performance perspective.

Right now, just instantiating a video.js player that loads the S3
URL (signed with a 1-hour expiration).

I'll also want to play around with setting up a cloudfront distribution
in front of it to see if that changes things. I think I can also
restrict based on IP address, so that's worth investigating.

I also plan on making an iframe-able version of this with a similar auth
scheme to surelink. Ie, a proof of concept with WC+S3 as the secure
streaming video server. If it seems workable, we can split it out into
its own service.

For now, this is just at a '/s3serve/' URL without anything linking to
it, so no one is likely to stumble on this unless we point them at
it. This should be OK to get started evaluating this approach and doing
some gap analysis against the h264/surelink setup.