retry operation on a DoesNotExist
if the view that created the operation didn't commit before calling
`.delay()` and the celery worker gets to it quickly, they `Operation`
object may not be in the database.

Ideally, the view would be careful to use
`@transaction.non_atomic_requests` and/or commit transactions
explicitly, but that's easy to miss.

We already have a retry with exponential backoff mechanism for
operations; this is actually the sort of thing that that should handle
quite well, so we might as well use it for this case as well.