return correct results when there are more than one File matching filter
see discussion on #108425

With the FLV conversions, it's now possible for a Video to have multiple
CUIT Files associated with it. The code had been assuming that the first
one returned by the database was the only one (previously perhaps not a
totally valid assumption, but almost always correct, so we didn't run
into it).

This should now work with multiple Files that match; returning the
expected value if any of them match the criteria.