meth flv convert command
manage command to run bulk flv -> mp4 conversions for mediathread
videos.

this seems like the simplest way to go about it. Just run the command
starting with a small number of conversions, check that they've run ok,
then run it again with a larger 'n'. Keep bumping it up until we run
into scaling issues or we've encoded all of the elligible videos.