extract metadata after upload to s3
this still has to happen on the flv -> mp4 conversions, especially if
the video is ever going to make it up to mediathread.

So, after uploading to s3, we just check whether there is already a
source file attached to the video. if not, we know we need to extract
metadata.