add a public/secure boolean field to collections
collections currently are all secure (videos uploaded to them are
uploaded to the secure directory) except one "h264 public" collection
that is basically hard-coded by title.

This has worked so far, but this really should be something that we can
set per collection. With the FLV import stuff now in particular, I have
a use-case for having separate collections for the flv imported videos,
one public and one secure.

So this is the first stage of supporting that. We just add the field to
the model. Once that migration has gone through the pipeline, we can
change the `is_public()` method to actually use it (and update
create/edit views).