Updates for Python 3
This makes a bunch of updates necessary for Python 3 compatibility.
There are still a bunch of unicode-related test errors in Python 3 -
these changes are just to make sure things are still working on Python
2.

librabbitmq needed to be updated from 1.x to 2.x. It doesn't look like
anything too big has changed here though:
https://github.com/celery/librabbitmq/blob/master/Changelog#L12

Because wardenclyffe has a lot going on, and I'm not the most familiar
with a lot of it, it may be best to confirm these changes are working on
staging before pushing to production. On my local wardenclyffe, I've
only confirmed that the basic Django stuff is working right.